Create default Minecraft Project setup. Rename SpigotConfig to DefaultProjectConfiguration Move DefaultProjectConfiguration to projects/config package.
