+++ src/main/java/net/senmori/project/config/DefaultProjectConfiguration.java
@@ -1,4 +1,4 @@
-package net.senmori.project.spigot.config;
+package net.senmori.project.config;
 
 import com.electronwill.nightconfig.core.CommentedConfig;
 import com.electronwill.nightconfig.core.file.FileNotFoundAction;
@@ -6,8 +6,6 @@ import com.electronwill.nightconfig.core.io.ConfigParser;
 import com.electronwill.nightconfig.core.io.ConfigWriter;
 import com.electronwill.nightconfig.core.io.ParsingMode;
 import net.senmori.project.asset.assets.ConfigurationFileAsset;
-import net.senmori.project.config.ConfigBuilder;
-import net.senmori.project.config.ConfigurationOptions;
 
 /**
  * A configuration object representing the Spigot project's configuration
@@ -16,16 +14,15 @@ import net.senmori.project.config.ConfigurationOptions;
  * This class should not be built directly; users should use
  * {@link ConfigBuilder} instead.
  */
-public class SpigotConfig {
+public class DefaultProjectConfiguration {
 
     private final ConfigurationOptions<CommentedConfig> configOptions;
-
     private final ConfigurationFileAsset configurationFileAsset;
 
     /**
-     * Use {@link ConfigBuilder} to create a {@link SpigotConfig} instance.
+     * Use {@link ConfigBuilder} to create a {@link DefaultProjectConfiguration} instance.
      */
-    public SpigotConfig(ConfigurationFileAsset configFileAsset, ConfigurationOptions<CommentedConfig> configuration, FileNotFoundAction action) {
+    public DefaultProjectConfiguration(ConfigurationFileAsset configFileAsset, ConfigurationOptions<CommentedConfig> configuration, FileNotFoundAction action) {
         this.configurationFileAsset = configFileAsset;
         this.configOptions = configuration;
         configOptions.getParser().parse(configFileAsset.getFile(), configOptions.getConfig(), ParsingMode.REPLACE, action);
