+++ src/main/resources/net/senmori/project/minecraft/mc_settings.toml
@@ -1,0 +1,2 @@
+[urls]
+version_manifest_link = "https://launchermeta.mojang.com/mc/game/version_manifest.json"
